troche gui, zmiana hasla, edycja usera

+++ src/dataModels/Uzytkownik.java
@@ -1,20 +1,18 @@
 package dataModels;
 
 public class Uzytkownik extends DataModel{
-  private Long id;
+  private Integer id;
   private String imie;
   private String nazwisko;
   private java.sql.Date data_urodzenia;
   private String login;
   private String telefon;
   private String haslo;
-  private Long poziom_uprawnien;
+  private Integer poziom_uprawnien;
 
-  public Long getId() {
-    return id;
-  }
+  public Integer getId() { return id; }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -66,11 +64,11 @@ public class Uzytkownik extends DataModel{
     this.haslo = haslo;
   }
 
-  public Long getPoziom_uprawnien() {
+  public Integer getPoziom_uprawnien() {
     return poziom_uprawnien;
   }
 
-  public void setPoziom_uprawnien(Long poziom_uprawnien) {
+  public void setPoziom_uprawnien(Integer poziom_uprawnien) {
     this.poziom_uprawnien = poziom_uprawnien;
   }
 }
